requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic-settings==2.0.3
requests==2.31.0
sqlalchemy==2.0.36
pymysql==1.1.0
cryptography==41.0.7
